requirements.txt
cffi==1.13.2
cryptography==2.8
Django==2.2.7
django-rest-auth==0.9.3
django-rest-knox==4.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
Pillow==6.2.1
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
